requirements.txt
numpy
scipy
pandas
matplotlib
pyyaml
pytables
xlrd
